[SC-30890] This image is now purescript-tools, similar to haskell-tools
We now build amd64 and arm64 versions of a purescript-tools image. We
use `bookworm` as the Debian version because the `trixie` unstable version
is not currently supported by nodesource.

The spago utility is being built from source, so that we have a version
that works on all of our developer arm64 machines. This utility does not
officially support arm64 so there may still be some issues with it.